requirements.txt
beautifulsoup4==4.14.2
faiss_cpu==1.12.0
fastapi==0.120.1
jsonlines==4.0.0
langchain_core==1.0.1
langchain_groq==1.0.0
numpy==2.3.4
pandas==2.3.3
pydantic==2.12.3
python-dotenv==1.2.1
Requests==2.32.5
sentence_transformers==5.1.2
